[ANDROID] Fix issue where already open apps would trigger dupe actions
This was the case even when `finish()` was called from userland. This
is because a warm app re-uses the activity object, resulting in cached
intents getting reprocessed during lifecycle events.

The fix was handling `onNewIntent` events, which allows us to capture
new activity and immediately clear the intent after first reading.